refactor: rename project readwith -> readum
- Rename base package com.readwith -> com.readum (28 files moved via git mv)
- Rename main/test class ReadwithApplication -> ReadumApplication
- Update settings.gradle rootProject.name, spring.application.name
- Update docker-compose container names, .env.example DB defaults
- Sync README.md and CLAUDE.md references
