+++ .env.example
@@ -1,6 +1,6 @@
 # MySQL
 MYSQL_ROOT_PASSWORD=your_root_password
-MYSQL_DATABASE=readwith
-MYSQL_USER=readwith
+MYSQL_DATABASE=readum
+MYSQL_USER=readum
 MYSQL_PASSWORD=your_password
 MYSQL_PORT=3306
