+++ README.md
@@ -1,4 +1,4 @@
-# readwith
+# readum
 
 ## 로컬 개발 환경 세팅
 
@@ -25,8 +25,8 @@ cp .env.example .env
 
 ```dotenv
 MYSQL_ROOT_PASSWORD=your_root_password   # MySQL root 계정 비밀번호
-MYSQL_DATABASE=readwith                  # 사용할 DB 이름
-MYSQL_USER=readwith                      # 애플리케이션 DB 유저
+MYSQL_DATABASE=readum                  # 사용할 DB 이름
+MYSQL_USER=readum                      # 애플리케이션 DB 유저
 MYSQL_PASSWORD=your_password             # 애플리케이션 DB 비밀번호
 MYSQL_PORT=3306                          # 호스트에 노출할 포트 (기본 3306, 충돌 시 변경)
 ```
@@ -65,7 +65,7 @@ docker-compose down -v   # 컨테이너 + 볼륨 삭제 (데이터 초기화)
 IntelliJ에서 Spring Boot를 `local` 프로파일로 실행하려면 Run Configuration을 아래와 같이 설정합니다.
 
 1. 상단 메뉴 **Run → Edit Configurations** 선택
-2. `ReadwithApplication` 실행 구성 선택 (없으면 `+` → `Spring Boot` 추가)
+2. `ReadumApplication` 실행 구성 선택 (없으면 `+` → `Spring Boot` 추가)
 3. 아래 항목을 설정합니다.
 
 **Active profiles**
@@ -77,7 +77,7 @@ local
 **Environment variables** — `.env`의 값을 그대로 입력합니다.
 
 ```
-MYSQL_ROOT_PASSWORD=your_root_password;MYSQL_DATABASE=readwith;MYSQL_USER=readwith;MYSQL_PASSWORD=your_password;MYSQL_PORT=3306
+MYSQL_ROOT_PASSWORD=your_root_password;MYSQL_DATABASE=readum;MYSQL_USER=readum;MYSQL_PASSWORD=your_password;MYSQL_PORT=3306
 ```
 
 > 세미콜론(`;`)으로 구분해 한 줄에 입력합니다.  
@@ -107,9 +107,9 @@ set -a && source .env && set +a && SPRING_PROFILES_ACTIVE=local ./gradlew bootRu
 | 변수명 | 설명 | 기본값 예시 |
 |--------|------|------------|
 | `MYSQL_ROOT_PASSWORD` | MySQL root 비밀번호 | `rootpassword` |
-| `MYSQL_DATABASE` | 사용할 데이터베이스 이름 | `readwith` |
-| `MYSQL_USER` | 애플리케이션 DB 유저 | `readwith` |
-| `MYSQL_PASSWORD` | 애플리케이션 DB 비밀번호 | `readwith1234` |
+| `MYSQL_DATABASE` | 사용할 데이터베이스 이름 | `readum` |
+| `MYSQL_USER` | 애플리케이션 DB 유저 | `readum` |
+| `MYSQL_PASSWORD` | 애플리케이션 DB 비밀번호 | `readum1234` |
 | `MYSQL_PORT` | 호스트에서 노출할 포트 | `3306` |
 
 ---
