+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'readwith'
+rootProject.name = 'readum'
