+++ src/main/java/com/readum/ReadumApplication.java
@@ -1,0 +1,13 @@
+package com.readum;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReadumApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ReadumApplication.class, args);
+	}
+
+}
