+++ src/main/java/com/readum/domain/example/dto/ExampleCreateCommand.java
@@ -1,0 +1,7 @@
+package com.readum.domain.example.dto;
+
+public record ExampleCreateCommand(
+        String name,
+        String description
+) {
+}
