+++ src/main/java/com/readum/domain/example/dto/ExampleCreateResult.java
@@ -1,0 +1,7 @@
+package com.readum.domain.example.dto;
+
+public record ExampleCreateResult(
+        Long id,
+        String name
+) {
+}
