+++ src/main/java/com/readum/domain/example/dto/ExampleResult.java
@@ -1,0 +1,22 @@
+package com.readum.domain.example.dto;
+
+import com.readum.model.example.entity.ExampleEntity;
+
+import java.time.LocalDateTime;
+
+public record ExampleResult(
+        Long id,
+        String name,
+        String description,
+        LocalDateTime createdAt
+) {
+
+    public static ExampleResult from(ExampleEntity entity) {
+        return new ExampleResult(
+                entity.getId(),
+                entity.getName(),
+                entity.getDescription(),
+                entity.getCreatedAt()
+        );
+    }
+}
