+++ src/main/java/com/readum/domain/exception/BadRequestException.java
@@ -1,0 +1,8 @@
+package com.readum.domain.exception;
+
+public class BadRequestException extends BusinessException {
+
+    public BadRequestException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
