+++ src/main/java/com/readum/domain/exception/ConflictException.java
@@ -1,0 +1,8 @@
+package com.readum.domain.exception;
+
+public class ConflictException extends BusinessException {
+
+    public ConflictException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
