+++ src/main/java/com/readum/domain/exception/ForbiddenException.java
@@ -1,0 +1,8 @@
+package com.readum.domain.exception;
+
+public class ForbiddenException extends BusinessException {
+
+    public ForbiddenException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
