+++ src/main/java/com/readum/domain/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package com.readum.domain.exception;
+
+public class NotFoundException extends BusinessException {
+
+    public NotFoundException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
