--- src/main/java/com/readwith/ReadwithApplication.java
@@ -1,13 +1,0 @@
-package com.readwith;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ReadwithApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ReadwithApplication.class, args);
-	}
-
-}
