--- src/main/java/com/readwith/domain/example/dto/ExampleCreateCommand.java
@@ -1,7 +1,0 @@
-package com.readwith.domain.example.dto;
-
-public record ExampleCreateCommand(
-        String name,
-        String description
-) {
-}
