--- src/main/java/com/readwith/domain/example/dto/ExampleCreateResult.java
@@ -1,7 +1,0 @@
-package com.readwith.domain.example.dto;
-
-public record ExampleCreateResult(
-        Long id,
-        String name
-) {
-}
