--- src/main/java/com/readwith/domain/exception/BadRequestException.java
@@ -1,8 +1,0 @@
-package com.readwith.domain.exception;
-
-public class BadRequestException extends BusinessException {
-
-    public BadRequestException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
