--- src/main/java/com/readwith/domain/exception/BusinessException.java
@@ -1,15 +1,0 @@
-package com.readwith.domain.exception;
-
-public class BusinessException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public BusinessException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
-    }
-}
