--- src/main/java/com/readwith/domain/exception/ConflictException.java
@@ -1,8 +1,0 @@
-package com.readwith.domain.exception;
-
-public class ConflictException extends BusinessException {
-
-    public ConflictException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
