--- src/main/java/com/readwith/domain/exception/ErrorCode.java
@@ -1,17 +1,0 @@
-package com.readwith.domain.exception;
-
-public enum ErrorCode {
-
-    // Example
-    EXAMPLE_NOT_FOUND("존재하지 않는 예시입니다.");
-
-    private final String message;
-
-    ErrorCode(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
