--- src/main/java/com/readwith/domain/exception/ForbiddenException.java
@@ -1,8 +1,0 @@
-package com.readwith.domain.exception;
-
-public class ForbiddenException extends BusinessException {
-
-    public ForbiddenException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
