--- src/main/java/com/readwith/domain/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package com.readwith.domain.exception;
-
-public class NotFoundException extends BusinessException {
-
-    public NotFoundException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
