--- src/main/java/com/readwith/presentation/common/ApiResponse.java
@@ -1,27 +1,0 @@
-package com.readwith.presentation.common;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public record ApiResponse<T>(T data, ErrorBody error) {
-
-    public record ErrorBody(String message) {}
-
-    public static <T> ResponseEntity<ApiResponse<T>> ok(T data) {
-        return ResponseEntity.ok(new ApiResponse<>(data, null));
-    }
-
-    public static <T> ResponseEntity<ApiResponse<T>> created(T data) {
-        return ResponseEntity
-                .status(HttpStatus.CREATED)
-                .body(new ApiResponse<>(data, null));
-    }
-
-    public static ResponseEntity<ApiResponse<?>> error(HttpStatus status, String message) {
-        return ResponseEntity
-                .status(status)
-                .body(new ApiResponse<>(null, new ErrorBody(message)));
-    }
-}
