+++ src/test/java/com/readum/ReadumApplicationTests.java
@@ -1,0 +1,13 @@
+package com.readum;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ReadumApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
