--- src/test/java/com/readwith/ReadwithApplicationTests.java
@@ -1,13 +1,0 @@
-package com.readwith;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReadwithApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
